requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.10
entrypoints==0.3
flask==3.0.3
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.11.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==8.5.0
incremental==16.10.1
itsdangerous==2.2.0
jinja2==3.1.4
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.13.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.3
oauthlib==3.1.0
opt-einsum==3.4.0
packaging==24.1
pexpect==4.6.0
pillow==10.4.0
powerline-status==2.7
protobuf==4.25.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygame==2.5.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
Twisted==18.9.0
typing-extensions==4.5.0
ubuntu-advantage-tools==27.11.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
zope.interface==4.7.1
